% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_forecast.R
\name{prep_forecast_data.point}
\alias{prep_forecast_data.point}
\title{Prepare point forecast data}
\usage{
prep_forecast_data.point(
  data,
  forecast_type = "point",
  observed_column,
  predicted_column,
  forecast_date,
  forecast_made,
  metric,
  other_characteristic_columns = NULL
)
}
\arguments{
\item{data}{A data frame containing the forecast data.}

\item{forecast_type}{The type of forecast ("point").}

\item{observed_column}{The name of the column containing observed values.}

\item{predicted_column}{The name of the column containing predicted values.}

\item{forecast_date}{The name of the column containing the forecast date.}

\item{forecast_made}{The name of the column containing the date the forecast was made.}

\item{metric}{The name of the metric being forecasted.}

\item{other_characteristic_columns}{Optional, other columns to be used as characteristics.}
}
\value{
A forecast_point object.
}
\description{
Prepare point forecast data
}
