% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_forecast.R
\name{prep_forecast_data.quantile}
\alias{prep_forecast_data.quantile}
\title{Prepare quantile forecast data}
\usage{
prep_forecast_data.quantile(
  data,
  forecast_type = "quantile",
  observed_column,
  predicted_column,
  quantile_columns,
  quantile_values,
  forecast_date,
  forecast_made,
  metric,
  other_characteristic_columns = NULL
)
}
\arguments{
\item{data}{A data frame containing the forecast data.}

\item{forecast_type}{The type of forecast ("quantile").}

\item{observed_column}{The name of the column containing observed values.}

\item{predicted_column}{The name of the column containing predicted values.}

\item{quantile_columns}{The name(s) of the column(s) containing quantiles (long format) or prefixes for quantile columns (wide format).}

\item{quantile_values}{Numeric values of the quantiles if the input is in wide format.}

\item{forecast_date}{The name of the column containing the forecast date.}

\item{forecast_made}{The name of the column containing the date the forecast was made.}

\item{metric}{The name of the metric being forecasted.}

\item{other_characteristic_columns}{Optional, other columns to be used as characteristics.}
}
\value{
A forecast_quantile object.
}
\description{
Prepare quantile forecast data
}
