% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_forecast.R
\name{prep_forecast_data}
\alias{prep_forecast_data}
\title{Prepare forecast data for scoring}
\usage{
prep_forecast_data(data, forecast_type, ...)
}
\arguments{
\item{data}{A data frame containing the forecast data.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A forecast object.
}
\description{
Prepare forecast data for scoring
}
