% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_forecast.R
\name{prep_forecast_data}
\alias{prep_forecast_data}
\title{Prepare forecast data for scoring}
\usage{
prep_forecast_data(data, forecast_type, ...)
}
\arguments{
\item{data}{A data frame containing the forecast data.}

\item{forecast_type}{The type of forecast ("point","quantile","sample").}

\item{...}{Further arguments passed to
\code{\link{prep_forecast_data.quantile}},
\code{\link{prep_forecast_data.point}},
\code{\link{prep_forecast_data.sample}}}
}
\value{
A forecast object.
}
\description{
Prepare forecast data for scoring using scoringutils::as_forecast,
which creates a forecast object (see ?scoringutils::as_forecast).
}
