% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_forecast.R
\name{prep_forecast_data}
\alias{prep_forecast_data}
\title{Prepare forecast data for scoring}
\usage{
prep_forecast_data(data, forecast_type, ...)
}
\arguments{
\item{data}{A data frame containing the forecast data.}

\item{forecast_type}{The type of forecast ("point","quantile","sample").}

\item{observed_column}{The name of the column containing observed values.}

\item{predicted_column}{The name of the column containing predicted values.}

\item{forecast_date}{The name of the column containing the forecast date.}

\item{forecast_made}{The name of the column containing the date the forecast was made.}

\item{metric}{The name of the metric being forecasted (e.g., cases, deaths, etc.).}

\item{other_characteristic_columns}{(optional) Other columns to be used as characteristics.}
}
\value{
A forecast object.
}
\description{
Prepare forecast data for scoring using scoringutils::as_forecast, which creates a forecast object (see ?scoringutils::as_forecast).
}
